navbar with categories

+++ src/app/components/cart/cart.component.ts
@@ -8,16 +8,21 @@ import { IProduct } from 'src/app/models/Product';
   styleUrls: ['./cart.component.css']
 })
 
-export class CartComponent{
+export class CartComponent implements OnInit{
 
   products: IProduct[] = []
 
   deleteProduct(id: number): void{
-    this.CartService.deleteItem(id)
-    this.products = this.CartService.getCart();
+    if(confirm('¿Seguro que desea eliminar el producto del carrito?')) {
+      this.CartService.deleteItem(id)
+      this.products = this.CartService.getCart();
+    }
+   
   }
 
-  constructor(private CartService: CartService) {
+  constructor(private CartService: CartService) {}
+
+  ngOnInit(){
     this.products = this.CartService.cart;
   }
